+++ src/dsp_tools/commands/excel2xml/excel2xml_lib.py
@@ -22,11 +22,11 @@ from dsp_tools.commands.excel2xml.propertyelement import PropertyElement
 from dsp_tools.error.custom_warnings import DspToolsUserWarning
 from dsp_tools.error.exceptions import BaseError
 from dsp_tools.legacy_models.datetimestamp import DateTimeStamp
-from dsp_tools.utils.date_formats.date_util import is_full_date
-from dsp_tools.utils.date_formats.shared import check_notna
-from dsp_tools.utils.date_formats.shared import simplify_name
-from dsp_tools.utils.date_formats.uri_util import is_iiif_uri
-from dsp_tools.utils.date_formats.uri_util import is_uri
+from dsp_tools.utils.data_formats.date_util import is_full_date
+from dsp_tools.utils.data_formats.shared import check_notna
+from dsp_tools.utils.data_formats.shared import simplify_name
+from dsp_tools.utils.data_formats.uri_util import is_iiif_uri
+from dsp_tools.utils.data_formats.uri_util import is_uri
 from dsp_tools.utils.xml_parsing.xml_schema_validation import parse_and_validate_xml_file
 
 # ruff: noqa: E501, UP031 (line-too-long, use f-string over percent formatting)
