+++ src/dsp_tools/commands/project/legacy_models/context.py
@@ -7,7 +7,7 @@ import regex
 from dsp_tools.commands.project.legacy_models.helpers import ContextType
 from dsp_tools.commands.project.legacy_models.helpers import OntoIri
 from dsp_tools.error.exceptions import BaseError
-from dsp_tools.utils.date_formats.iri_util import is_iri
+from dsp_tools.utils.data_formats.iri_util import is_iri
 
 
 class ContextIterator:
