+++ src/dsp_tools/commands/xmlupload/prepare_xml_input/transform_input_values.py
@@ -13,8 +13,8 @@ from dsp_tools.commands.xmlupload.models.intermediary.values import Intermediary
 from dsp_tools.commands.xmlupload.models.intermediary.values import IntermediaryValueTypes
 from dsp_tools.commands.xmlupload.models.intermediary.values import IntervalFloats
 from dsp_tools.error.exceptions import InputError
-from dsp_tools.utils.date_formats.date_util import Date
-from dsp_tools.utils.date_formats.date_util import parse_date_string
+from dsp_tools.utils.data_formats.date_util import Date
+from dsp_tools.utils.data_formats.date_util import parse_date_string
 
 InputTypes: TypeAlias = Union[str, FormattedTextValue]
