+++ src/dsp_tools/utils/data_formats/shared.py
