+++ test/unittests/commands/xmlupload/make_rdf_graph/test_make_values.py
@@ -33,10 +33,10 @@ from dsp_tools.commands.xmlupload.models.lookup_models import JSONLDContext
 from dsp_tools.commands.xmlupload.models.permission import Permissions
 from dsp_tools.commands.xmlupload.models.permission import PermissionValue
 from dsp_tools.error.exceptions import BaseError
-from dsp_tools.utils.date_formats.date_util import Calendar
-from dsp_tools.utils.date_formats.date_util import Date
-from dsp_tools.utils.date_formats.date_util import Era
-from dsp_tools.utils.date_formats.date_util import SingleDate
+from dsp_tools.utils.data_formats.date_util import Calendar
+from dsp_tools.utils.data_formats.date_util import Date
+from dsp_tools.utils.data_formats.date_util import Era
+from dsp_tools.utils.data_formats.date_util import SingleDate
 
 ONTO = Namespace("http://0.0.0.0:3333/ontology/9999/onto/v2#")
